Work in Progress
This is a table-view row that allows the user to enter data in the same form
as the data is displayed in the table. It uses some client-side JavaScript to
allow the user to enter an arbitrary number of rows of data at the same time.

Initial check-in. There's enough code, special cases and workarounds here that
I want the file under version control. This still needs some work -- for
example, processing is not working at all. There are several related changes to
other parts of Swat that I am not checking in at this point.


svn commit r2924