New class SwatDetailsStore.
Parallel to SwatTableStore, but for a details view instead of a table store.  Provides more structure to the data to be displayed in the details view and can be based on another object.  In practice we can base a details store on a $row object or a dataobject and modify and add fields as necessary for the view.  Really, it works well.


git-svn-id: file:///Users/gauthierm/svnrepo/swat/trunk@3338 94490950-a86c-4493-9ccb-530f72933e47