Add SwatDB::exec() method parallel to MDB2::exec() method.
MDB2::query() no longer returns the number of affected rows when there is no recordset to return.  Intsead MDB now offers MDB2::exec() to run an SQL statement that does not produce a recordset.  It will always return the number of rows affected.

So what does this mean?
Places we were using SwatDB::query() and expecting the number of affected rows returned need to be changed to SwatDB:exec().
Places we were using SwatDB::query() and ignoring the return value are fine to stay as SwatDB:query(), but in future code our covention should be to use SwatDB::exec().



svn commit r3425