Add new SwatUIObject::copy() method that performs an explicit deep copy of the UI tree with an optional id prefix to prepend to all copied UI object ids.
This isn't perfect yet but I'm commiting it because it works better than what existed before and because it's a large change.

The current limitation is it only works on uninit'd UI trees. This is because of additional references outside the UI tree that are hard to copy properly. These references get automatically assigned during init() for the most part. Culprits are view selectors, input cells and replicated widgets within replicable objects. If your UI does not contain views or replicable objects it should be ok to use copy() on an init'd tree.

I intend to remove the above limitation at some point in the future.



git-svn-id: file:///Users/gauthierm/svnrepo/swat/trunk@22984 94490950-a86c-4493-9ccb-530f72933e47