Thought I fixed this way back in r68409, but turns out I didn't.
With the old code, if $end_year - $start_year was 1, but the valid range was longer than a year, the month flydown would not display all the valid months.
So, if the valid_start was Dec 1 2011 and the valid_end was Dec 30 2012, we'd only display december in the month flydown. Now we make sure the difference is greater than a year before displaying the limited month list. 



svn commit r77354