Add getAttributes() method to SwatDBDataObject
Now that we're using getter and setter patterns in our dataobjects we
can no longer use get_object_vars() on the dataobject to get all the
properties we care about, for example when editing with admin tools.

getAttributes() is a replacement for places where we have relied on
get_object_vars() in the past to remedy this.