Add two checkbox trees that track dependencies
Two new checkbox tree widgets that keep their update child and parent
state when an option is checked or unchecked.

The first type is the parent dependency tree which ensures that all
parent options are checked when a child option is selected.

The second type is the child dependency tree which ensures that all
child options are checked when a parent option is selected.