Allow for list items with no checkbox to be walked
Before this change any item that didn’t have a checkbox would be
completely ignored by the walk function. As a result the function
returned by walk would not be able to manipulate that item or any of
that item’s decedents.

When a node with no checkbox was at the top of the tree all of that
node's children weren’t being walked over. This meant that widgets that
used the function built by walk would not work as expected. For example
the check all widget was not checking any item that lived below a node
with no checkbox. Instead of ignoring these items we now allow them to
be chained like regular nodes.

The previous fix selected all decedents of a node instead of the direct
children. This would cause the same nodes to be walked over and over
and over again which worked but is both inefficient and incongruent with
the rest of the code in SwatCheckboxTree. The walk function is recursive
so these decedents will be reached by subsequent recursive calls.