SwatMessage Fix set primary_content / sprintf error
Fix to prevent sprintf function called with a primary message string which may
have a substring that can be unintentionally parsed as an sprintf format
specifier (for example numerous promo codes have % followed by characters that will cause a
parsing error).

Remove code that attempts (and fails) to append, in a SwatTitleable case, the
title string to the primary_content string. Calling code exists that expects a
primary_content value addMessage() to not add anything to the primary_content
value.