Fix invalid locale parameter to IntlDateFormatter
Since `Fixed bug GH-12282` in php 8.1.25 exceptions are thrown when passing an
invalid locale string to IntlDateFormatter().

The setlocale(LC_TIME, 0) function call returns string `C` (the contextless
locale OS default) which is invalid in php's IntlDateFormatter method.

The first parameter of IntlDateFormatter (which is $local) can be null, and if
null there's no exception and the locale appears to be `en` by default which is
the desired behavour.

Removing block that reassigned $locale to `C` if $local parameter is  null.